+++ R/light_interaction.R
@@ -1,10 +1,10 @@
 #' Interaction Strength
 #'
 #' Based on (zero) mean-centered ICE curves, different measures of interaction strength are calculated.
-#' For the default (\code{type = "overall"}), we get a single value per input variable z.
+#' For the default (\code{pairwise = FALSE}), we get a single value per input variable z.
 #' It measures the variance in c-ICE curves explained by all interactions of z
 #' (i.e. the variance of all effects of z unexplained by the main effect of z).
-#' For \code{type = "pairwise"}, for each combination of variable pairs z1, z2, a value is provided.
+#' For \code{pairwise = TRUE}, for each combination of variable pairs z1, z2, a value is provided.
 #' It measures the variance in c-ICE curves explained by the pairwise interaction terms and is
 #' assessed by calculating the squared difference between two-dimensional ICE curves spanned by
 #' z1 and z2 and the two corresponding one-dimensional ICE curves. The values can be normalized by the total variance.
@@ -12,18 +12,18 @@
 #' while normalized values show interaction strength relative to the overall effects of z (resp. z1 and z2).
 #' In order to simplify interpretation, unnormalized values are root-transformed to be on the scale of the response variable.
 #' Normalized values can be interpreted as proportions of variance explained. In order to reproduce Friedman's H statistic [1],
-#' choose \code{type = "pairwise"}, \code{normalize = TRUE} and \code{take_sqrt = TRUE}.
+#' choose \code{pairwise = TRUE}, \code{normalize = TRUE} and \code{take_sqrt = TRUE}.
 #'
 #' Note that continuous variables are binned using quantile cuts to get more stable results. The minimum required elements in the (multi-)flashlight are "predict_function", "model", "linkinv" and "data", where the latest can be passed on the fly.
 #'
 #' @importFrom stats setNames
-#' @importFrom dplyr is.tbl as_tibble bind_rows bind_cols
+#' @importFrom dplyr is.tbl as_tibble bind_rows bind_cols arrange_at expand_grid
 #' @importFrom utils combn
 #' @param x An object of class \code{flashlight} or \code{multiflashlight}.
 #' @param data An optional \code{data.frame}.
 #' @param by An optional vector of column names used to additionally group the results.
 #' @param v Vector of variables to be assessed.
-#' @param type Type of the measure of interaction strength: Either "overall" or "pairwise".
+#' @param pairwise Should overall interaction strength per variable be shown or pairwise interactions? Defaults to \code{FALSE}.
 #' @param normalize Should the variances explained be normalized? Default is \code{FALSE}.
 #' @param take_sqrt By default, unnormalized values are root-transformed while normalized values are not.
 #' @param grid_size Grid size used to form the outer product. Will be randomly picked from data.
@@ -54,7 +54,7 @@
 #' fls <- multiflashlight(list(fl_additive, fl_nonadditive), data = iris, y = "Sepal.Length")
 #' x <- fls$nonadditive
 #' plot(light_interaction(fls))
-#' plot(light_interaction(fls, type = "pairwise", normalize = TRUE))
+#' plot(light_interaction(fls, pairwise = TRUE, normalize = TRUE))
 #' plot(light_interaction(fls, by = "Species"), swap_dim = TRUE)
 #' @seealso \code{\link{light_ice}}.
 light_interaction <- function(x, ...) {
@@ -70,20 +70,17 @@ light_interaction.default <- function(x, ...) {
 #' @describeIn light_interaction Interaction strengths for a flashlight object.
 #' @export
 light_interaction.flashlight <- function(x, data = x$data, by = x$by,
-                                         v = NULL, type = c("overall", "pairwise"),
+                                         v = NULL, pairwise = FALSE,
                                          normalize = FALSE,
                                          take_sqrt = !normalize,
-                                         grid_size = 20,
-                                         n_max = 200, seed = NULL,
-                                         use_linkinv = FALSE,
+                                         grid_size = 30, n_max = 200,
+                                         seed = NULL, use_linkinv = FALSE,
                                          value_name = "value",
                                          error_name = "error", label_name = "label",
                                          variable_name = "variable", ...) {
-  type <- match.arg(type)
-
   stopifnot((n <- nrow(data)) >= 1L,
             !anyDuplicated(c(by, v, value_name, label_name, error_name, variable_name,
-                             "id_", "value_", "value_i", "value_j", "denominator_")))
+                             "id_", "id_curve", "value_", "value_i", "value_j", "denom_")))
 
   if (is.null(v)) {
     v <- setdiff(colnames(data), c(x$y, by))
@@ -100,14 +97,14 @@ light_interaction.flashlight <- function(x, data = x$data, by = x$by,
 
   # Select grid indices from the reduced data set
   if (grid_size < n) {
-    grid_ind <- sample(n, grid_size)
+    grid_id <- sample(n, grid_size)
   } else {
     grid_size <- n
-    grid_ind <- seq_len(grid_size)
+    grid_id <- seq_len(grid_size)
   }
 
   # Update flashlight
-  x <- flashlight(x, data = data, by = by,
+  x <- flashlight(x, by = by, data = data,
                   linkinv = if (use_linkinv) x$linkinv else function(z) z)
 
   # Function to safely replace bad values by 0
@@ -119,46 +116,68 @@ light_interaction.flashlight <- function(x, data = x$data, by = x$by,
     x
   }
 
-  # ICE wrapper used in the core function
-  call_ice <- function(grid, vn, drop_w = FALSE) {
-    ice <- light_ice(x, grid = grid, n_max = Inf, value_name = vn,
-                     id_name = "id_")$data
-    ice[, setdiff(colnames(ice), c(label_name, if (drop_w) x$w))]
+  # Simplified version of light_profile and light_ice
+  call_pd <- function(z, X = data, vn = "value_", only_values = FALSE, agg = TRUE) {
+    cols <- colnames(X)
+    if (!is.null(x$w)) {
+      ww <- X[grid_id, x$w, drop = FALSE]
+      ww[["id_"]] <- grid_id
+    }
+    grid <- X[grid_id, z, drop = FALSE]
+    grid[["id_"]] <- grid_id
+    X[, z] <- NULL
+    X[["id_curve"]] <- seq_len(nrow(X))
+    X <- expand_grid(X, grid)
+    X[[vn]] <- predict(x, data = X[, cols, drop = FALSE])
+    if (!agg) {
+      X[[vn]] <- grouped_center(X, x = vn, w = x$w, by = "id_curve")
+      return(X)
+    }
+    out <- grouped_stats(X, vn, w = x$w, by = c("id_", by), counts = FALSE, na.rm = TRUE)
+    out <- arrange_at(out, c("id_", by))
+    if (!is.null(x$w)) {
+      out[[x$w]] <- ww[match(out[["id_"]], ww[["id_"]]), x$w]
+    }
+    out[[vn]] <- grouped_center(out, x = vn, w = x$w, by = by)
+    if (only_values) out[, vn, drop = FALSE] else out
   }
+
+  # Functions that calculates the test statistic by variable (pair)
   core_func <- function(z) {
-    # Generate data and the values of the numerator
-    if (type == "overall") {
-      dat <- call_ice(data[grid_ind, z, drop = FALSE], "value_")
-      dat[[value_name]] <- grouped_center(dat, x = "value_", w = x$w, by = c(by, z))^2
-    } else {
-      dat_ij <- call_ice(data[grid_ind, z], "value_")
-      dat_i <- call_ice(data[grid_ind, z[1], drop = FALSE], "value_i", drop_w = TRUE)
-      dat_j <- call_ice(data[grid_ind, z[2], drop = FALSE], "value_j", drop_w = TRUE)
-      dat <- bind_cols(dat_ij, dat_i, dat_j)
-      dat[, c("value_j", "value_i", "value_")] <- scale(dat[, c("value_j", "value_i", "value_")], scale = F)
+    if (pairwise) {
+      dat <- bind_cols(call_pd(z),
+                       call_pd(z[1], vn = "value_i", only_values = TRUE),
+                       call_pd(z[2], vn = "value_j", only_values = TRUE))
       dat[[value_name]] <- (dat[["value_"]] - dat[["value_i"]] - dat[["value_j"]])^2
     }
-
-    # Aggregate numerator by groups
-    numerator <- grouped_stats(dat, x = value_name, w = x$w, by = by,
-                               counts = FALSE, na.rm = TRUE)
-    if (!normalize) {
-      return(numerator)
+    else {
+      dat <- call_pd(z, agg = FALSE)
+      dat[[value_name]] <- grouped_center(dat, x = "value_", w = x$w, by = c(by, z))^2
     }
 
-    dat[["denominator_"]] <- dat[["value_"]]^2
-    denominator <- grouped_stats(dat, x = "denominator_", w = x$w, by = by,
+    # Aggregate num by groups
+    num <- grouped_stats(dat, x = value_name, w = x$w, by = by,
+                               counts = FALSE, na.rm = TRUE)
+    if (!normalize) {
+      return(num)
+    }
+
+    # Calculate denom by groups
+    dat[["denom_"]] <- dat[["value_"]]^2
+    denom <- grouped_stats(dat, x = "denom_", w = x$w, by = by,
                                  counts = FALSE, na.rm = TRUE)
 
     # Calculate ratio
-    numerator[[value_name]] <- replace_bad(numerator[[value_name]]) / denominator[["denominator_"]]
-    numerator
+    num[[value_name]] <- replace_bad(num[[value_name]]) / denom[["denom_"]]
+    num
   }
-  if (type == "overall") {
-    data <- setNames(lapply(v, core_func), v)
-  } else {
+
+  # Calculate statistic for each variable (pair) and combine results
+  if (pairwise) {
     v_pairs <- combn(v, 2, simplify = FALSE)
     data <- setNames(lapply(v_pairs, core_func), lapply(v_pairs, paste, collapse = ":"))
+  } else {
+    data <- setNames(lapply(v, core_func), v)
   }
   data <- bind_rows(data, .id = variable_name)
   data[[value_name]] <- replace_bad(data[[value_name]])
@@ -183,6 +202,7 @@ light_interaction.flashlight <- function(x, data = x$data, by = x$by,
   class(out) <- c("light_importance", "light", "list")
   out
 }
+
 
 #' @describeIn light_interaction for a multiflashlight object.
 #' @export